move models into lib
