+++ server/lib/db/transactions.js
@@ -1,4 +1,4 @@
-const Transactions = require('../../models/transaction.js');
+const Transactions = require('../models/transaction.js');
 const config       = require('../../config');
 const logger       = require('../logger');
