+++ server/lib/models/block.js
@@ -1,6 +1,6 @@
 const mongoose = require('mongoose');
-const config = require('../config');
-const util = require('../lib/util');
+const config = require('../../config');
+const util = require('../util');
 
 const Schema = mongoose.Schema;
 // These limits can be overriden higher up the stack
