+++ server/test/model.test.js
@@ -1,5 +1,5 @@
 const db = require('../lib/db');
-const Block = require('../models/block.js');
+const Block = require('../lib/models/block.js');
 
 Block.findOne({}, (err, block) => {
   console.log(err);
